refactor: move files around, tweak api (wip)
