+++ packages/openapi-parser/package.json
@@ -53,6 +53,7 @@
     "jsonpointer": "^5.0.1",
     "leven": "^4.0.0",
     "openapi-types": "^12.1.3",
-    "vite": "^5.1.4"
+    "vite": "^5.1.4",
+    "yaml": "^2.3.4"
   }
 }
