+++ packages/openapi-parser/src/utils/toJson.ts
@@ -1,0 +1,2 @@
+export const toJson = (value: Record<string, any>) =>
+  JSON.stringify(value, null, 2)
