+++ packages/openapi-parser/src/utils/toYaml.ts
@@ -1,0 +1,3 @@
+import YAML from 'yaml'
+
+export const toYaml = (value: Record<string, any>) => YAML.stringify(value)
